+++ gnparser/cmd/root.go
@@ -147,7 +147,7 @@ func versionFlag(cmd *cobra.Command) {
 		os.Exit(1)
 	}
 	if version {
-		var gnp *gnparser.GNparser
+		gnp := gnparser.NewGNparser()
 		fmt.Printf("\nversion: %s\n\nbuild:   %s\n\n",
 			gnp.Version(), gnp.Build())
 		os.Exit(0)